[STRUCTURAL] Extract Class: subagent wiring into internal/subagents
Phase 3 slice 5. The subagent system — the agent definition registry with
its built-in and config-declared entries, the wake manager, the spawner,
the worktree provider, and the task/list_tasks tools — was written twice,
once in runInteractive and once in runHeadless, ~55 lines each. It is
policy rather than plumbing: which definitions ship built in, how config
declarations are translated, and whether a subagent gets its own
worktree.

The two copies had already diverged in three ways, so a faithful move has
to parameterise them rather than pick a winner. Options carries all
three: WorktreeManager (interactive passed its session manager, headless
did not), Logf (headless logged registration failures, interactive
discarded them), and the enable flags (headless skipped construction
entirely when both tools were off, avoiding a git call). Each call site
keeps exactly the behaviour it had; the divergences are fixed in
follow-up commits so the fixes are reviewable as behaviour changes.

The five adapters move with it — they had no other callers — as do their
seven tests. gitRepoRoot names the repository-root lookup both modes
open-coded.

The moved worktree adapter test built a manager over this checkout; it
now builds a throwaway repository, which is the trap slice 4 hit.

main.go: 2,795 -> 2,621 lines.
